fix(frontend): resolve lint violations and refresh-safe exports
