+++ frontend/src/components/nav-main.tsx
@@ -10,9 +10,9 @@ import {
   SidebarMenuItem,
   SidebarMenuSub,
   SidebarMenuSubButton,
-  SidebarMenuSubItem,
-  useSidebar,
+  SidebarMenuSubItem
 } from "@/components/ui/sidebar"
+import { useSidebar } from "@/components/ui/sidebar-context"
 
 export function NavMain({
   items,
