+++ frontend/src/components/ui/badge.tsx
@@ -49,4 +49,4 @@ function Badge({
   })
 }
 
-export { Badge, badgeVariants }
+export { Badge }
